fix: added handling of packagemanagermanifest asset
Some Unity packages have a `packagemanagermanifest` asset in them, with no pathname or anything else other than the `asset` file. I'm almost certain this is a `Packages/manifest.json` file that contains the dependencies of the package, since installing this package in Unity will prompt you to add those dependencies. This commit handle those case so the extractor doesn't crash when encountering this file.

Also added a way to keep the extracted files.